pyproject.toml
[tool.poetry]
name = "gcs-metric-extract"
version = "0.1.0"
description = "A little help getting GCS metrics out of GCP"

license = "Apache 2.0"

authors = [
    "Dom Zippilli <dzippilli@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.10" 
click = "^8.1.3"
google-cloud-monitoring = "^2.10.1"

[tool.poetry.scripts]
gcs_metric_extract = "gcs_metric_extract:main"

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 4
split_before_logical_operator = true